backend/python/requirements.txt
pydicom>=2.4.0
numpy>=1.21.0
pillow>=9.0.0